refactor(app): share status drift filtering
