+++ openspec/changes/refactor-app-status-drift-filter/README.md
@@ -1,0 +1,3 @@
+# refactor-app-status-drift-filter
+
+Refactor: share status drift filtering between CLI and MCP
