+++ src/mcp/tools.rs
@@ -9,7 +9,7 @@ use rmcp::{
 
 use crate::app::next_actions::{next_action_code, ordered_next_actions};
 use crate::app::operator_assets::{check_operator_command_dir, check_operator_file};
-use crate::app::status_drift::{drift_summary, drift_summary_by_root};
+use crate::app::status_drift::{drift_summary_by_root, filter_drift_by_kind};
 use crate::user_error::UserError;
 
 use super::confirm::{CONFIRM_TOKEN_TTL, ConfirmTokenBinding};
@@ -606,17 +606,8 @@ async fn call_status_in_process(args: StatusArgs) -> anyhow::Result<(String, ser
                 })
                 .collect();
 
-            let (drift, summary, summary_total_opt) = if only_kinds.is_empty() {
-                (drift, summary_total, None)
-            } else {
-                let drift: Vec<crate::handlers::status::DriftItem> = drift
-                    .into_iter()
-                    .filter(|d| only_kinds.contains(d.kind.as_str()))
-                    .collect();
-                let summary = drift_summary(&drift);
-
-                (drift, summary, Some(summary_total))
-            };
+            let (drift, summary, summary_total_opt) =
+                filter_drift_by_kind(drift, &only_kinds, summary_total);
             let summary_by_root = drift_summary_by_root(&drift);
 
             let mut data = serde_json::json!({
